chore(pmtud/tcp): move test helpers in helpers_test.go
